finish aht20 implementation

+++ BtHomeSensor/Nordic/nrf52840/app/library/aht20/include/aht20Interface.hpp
@@ -5,8 +5,8 @@ class aht20Interface
 
 public:
   virtual int8_t Init() = 0;
-  virtual uint32_t ReadHumidity(void) = 0;
-  virtual uint32_t ReadTemperature(void) = 0;
+  virtual float ReadHumidity(void) = 0;
+  virtual float ReadTemperature(void) = 0;
   virtual int8_t TriggerMeasurement(void (*delayFunc)(uint32_t), uint32_t ms) = 0;
   ;
 };
